requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
stripe
requests
email-validator
python-dotenv
Flask-Migrate